chore(Scripts/SethekkHalls): Move Anzu script to its own file
